v9.0.3 发布

+++ src/main/java/com/ujcms/cms/core/service/DictTypeService.java
@@ -1,13 +1,16 @@
 package com.ujcms.cms.core.service;
 
-import com.github.pagehelper.PageHelper;
+import com.github.pagehelper.page.PageMethod;
+import com.ujcms.cms.core.domain.Dict;
 import com.ujcms.cms.core.domain.DictType;
+import com.ujcms.cms.core.domain.base.DictTypeBase;
 import com.ujcms.cms.core.listener.SiteDeleteListener;
 import com.ujcms.cms.core.mapper.DictMapper;
 import com.ujcms.cms.core.mapper.DictTypeMapper;
 import com.ujcms.cms.core.service.args.DictTypeArgs;
 import com.ujcms.commons.query.QueryInfo;
 import com.ujcms.commons.query.QueryParser;
+import com.ujcms.commons.web.exception.LogicException;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,22 +27,24 @@ import java.util.Objects;
 public class DictTypeService implements SiteDeleteListener {
     private final DictTypeMapper mapper;
     private final DictMapper dictMapper;
+    private final DictService dictService;
     private final SeqService seqService;
 
-    public DictTypeService(DictTypeMapper mapper, DictMapper dictMapper, SeqService seqService) {
+    public DictTypeService(DictTypeMapper mapper, DictMapper dictMapper,
+                           DictService dictService, SeqService seqService) {
         this.mapper = mapper;
         this.dictMapper = dictMapper;
+        this.dictService = dictService;
         this.seqService = seqService;
     }
 
     @Transactional(rollbackFor = Exception.class)
     public void insert(DictType bean, Integer siteId) {
-        bean.setId(seqService.getNextVal(DictType.TABLE_NAME));
+        bean.setId(seqService.getNextVal(DictTypeBase.TABLE_NAME));
         // 全局共享数据的站点ID设置为null
         bean.setSiteId(bean.isGlobal() ? null : siteId);
         mapper.insert(bean);
     }
-
 
     @Transactional(rollbackFor = Exception.class)
     public void update(DictType bean, Integer siteId) {
@@ -60,12 +65,30 @@ public class DictTypeService implements SiteDeleteListener {
 
     @Transactional(rollbackFor = Exception.class)
     public int delete(Integer id) {
+        if (dictMapper.existsByTypeId(id) > 0) {
+            throw new LogicException("error.refer.dict");
+        }
         return mapper.delete(id);
     }
 
     @Transactional(rollbackFor = Exception.class)
     public int delete(List<Integer> ids) {
         return ids.stream().filter(Objects::nonNull).mapToInt(this::delete).sum();
+    }
+
+    @Transactional(rollbackFor = Exception.class)
+    public void copyBySite(Integer siteId, Integer fromSiteId) {
+        List<DictType> list = listBySiteId(fromSiteId);
+        for (DictType type : list) {
+            Integer origTypeId = type.getId();
+            insert(type, siteId);
+            for (Dict dict : dictService.listByTypeId(origTypeId)) {
+                dict.setTypeId(type.getId());
+                // 暂不支持上下级结构
+                dict.setParentId(null);
+                dictService.insert(dict);
+            }
+        }
     }
 
     @Nullable
@@ -75,17 +98,20 @@ public class DictTypeService implements SiteDeleteListener {
 
 
     public List<DictType> selectList(DictTypeArgs args) {
-        QueryInfo queryInfo = QueryParser.parse(args.getQueryMap(), DictType.TABLE_NAME, "order,id");
+        QueryInfo queryInfo = QueryParser.parse(args.getQueryMap(), DictTypeBase.TABLE_NAME, "order,id");
         return mapper.selectAll(queryInfo);
     }
 
     public List<DictType> selectList(DictTypeArgs args, int offset, int limit) {
-        return PageHelper.offsetPage(offset, limit, false).doSelectPage(() -> selectList(args));
+        return PageMethod.offsetPage(offset, limit, false).doSelectPage(() -> selectList(args));
+    }
+
+    public List<DictType> listBySiteId(Integer siteId) {
+        return selectList(DictTypeArgs.of().siteId(siteId));
     }
 
     public boolean existsByAlias(String alias, @Nullable Integer siteId) {
-        return PageHelper.offsetPage(0, 1, false).<Number>doSelectPage(() ->
-                mapper.countByAlias(alias, siteId)).iterator().next().intValue() > 0;
+        return mapper.existsByAlias(alias, siteId) > 0;
     }
 
     @Override
